feat: 이미지 DB 파일 경로 컬럼 추가
