+++ src/main/java/com/three/ott_suggestion/image/repository/PostImageRepository.java
@@ -1,7 +1,6 @@
 package com.three.ott_suggestion.image.repository;
 
-import com.three.ott_suggestion.image.PostImage;
-import java.util.Optional;
+import com.three.ott_suggestion.image.entity.PostImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostImageRepository extends JpaRepository<PostImage, Long> {
