+++ src/main/java/com/three/ott_suggestion/image/repository/UserImageRepository.java
@@ -1,7 +1,6 @@
 package com.three.ott_suggestion.image.repository;
 
-import com.three.ott_suggestion.image.UserImage;
-import java.util.Optional;
+import com.three.ott_suggestion.image.entity.UserImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserImageRepository extends JpaRepository<UserImage,Long> {
